// FrontentTravelAgency/src/app/model/Destination.ts
export class Destination {
  image: any | undefined; // o sa aflu ce se intampla cu asta
  id: number | undefined;
  name: string | undefined;
  description: string | undefined;
  location: string | undefined;
  numberOfRooms: number | undefined;
  numberOfSeatsAvailable: number | undefined;
  numberOfSeatsTotal: number | undefined;
  price: number | undefined;
  offer: number | undefined;
  editable: boolean | undefined;
}